flask_app/requirements.txt
argcomplete==1.9.3
certifi==2019.9.11
cfn-flip==1.2.1
chardet==3.0.4
click==7.0
docutils==0.15.2
durationpy==0.5
flask==1.1.1
flask-oidc==1.4.0
flask-sqlalchemy==2.3.2
future==0.16.0
hjson==3.0.1
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.1
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
markupsafe==1.1.1
oauth2client==4.1.3
okta
placebo==0.9.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dateutil==2.6.1
python-slugify==1.2.4
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
sqlalchemy==1.3.8
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.1
unidecode==1.1.1
urllib3==1.25.6
werkzeug==0.16.0
wsgi-request-logger==0.4.6

